requirements.txt
ultralytics == 8.2.21
streamlit == 1.35.0
pytesseract == 0.3.10
pillow == 10.3.0
opencv-python == 4.9.0.80